# packages/shared-types/openapi.json
{
  "components": {
    "schemas": {
      "ApiEnvelope_GameData_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/GameData"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "ApiEnvelope[GameData]",
        "type": "object"
      },
...
      "ApiEnvelope_WeatherData_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/WeatherData"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "ApiEnvelope[WeatherData]",
        "type": "object"
      },
...
      "ApiEnvelope_list_GameData__": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/GameData"
            },
            "title": "Data",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "ApiEnvelope[list[GameData]]",
        "type": "object"
      },
      "ApiEnvelope_list_WeatherData__": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/WeatherData"
            },
            "title": "Data",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "ApiEnvelope[list[WeatherData]]",
        "type": "object"
      },
      "ApiMeta": {
        "additionalProperties": false,
        "properties": {
          "data_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data Version"
          },
          "pagination": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PaginationMeta"
              },
              {
                "type": "null"
              }
            ]
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "source_time": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Time"
          },
          "stale": {
            "default": false,
            "title": "Stale",
            "type": "boolean"
          }
        },
        "required": [
          "request_id"
        ],
        "title": "ApiMeta",
        "type": "object"
      },
...
      "ErrorBody": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorEnvelope": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
      "GameData": {
        "additionalProperties": false,
        "properties": {
          "away_probable_pitcher": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PitcherBrief"
              },
              {
                "type": "null"
              }
            ]
          },
          "away_team": {
            "$ref": "#/components/schemas/TeamBrief"
          },
          "game_date": {
            "format": "date",
            "title": "Game Date",
            "type": "string"
          },
          "game_id": {
            "title": "Game Id",
            "type": "string"
          },
          "game_time_utc": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Game Time Utc"
          },
          "home_probable_pitcher": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PitcherBrief"
              },
              {
                "type": "null"
              }
            ]
          },
          "home_team": {
            "$ref": "#/components/schemas/TeamBrief"
          },
          "season": {
            "title": "Season",
            "type": "integer"
          },
          "source_updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Updated At"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "venue": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VenueBrief"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "game_id",
          "game_date",
          "season",
          "away_team",
          "home_team"
        ],
        "title": "GameData",
        "type": "object"
      },
...
      "PaginationMeta": {
        "additionalProperties": false,
        "properties": {
          "limit": {
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "limit"
        ],
        "title": "PaginationMeta",
        "type": "object"
      },
      "PitcherBrief": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "player_id": {
            "title": "Player Id",
            "type": "string"
          }
        },
        "required": [
          "player_id"
        ],
        "title": "PitcherBrief",
        "type": "object"
      },
...
      "TeamBrief": {
        "additionalProperties": false,
        "properties": {
          "abbreviation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Abbreviation"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          }
        },
        "required": [
          "name"
        ],
        "title": "TeamBrief",
        "type": "object"
      },
      "VenueBrief": {
        "additionalProperties": false,
        "properties": {
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "roof_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Roof Type"
          }
        },
        "required": [
          "name"
        ],
        "title": "VenueBrief",
        "type": "object"
      },
...
      "WeatherData": {
        "additionalProperties": false,
        "properties": {
          "available": {
            "title": "Available",
            "type": "boolean"
          },
          "away_team": {
            "$ref": "#/components/schemas/TeamBrief"
          },
          "condition": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Condition"
          },
          "edge_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Edge Label"
          },
          "feels_like_f": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feels Like F"
          },
          "forecast_for": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Forecast For"
          },
          "game_date": {
            "format": "date",
            "title": "Game Date",
            "type": "string"
          },
          "game_id": {
            "title": "Game Id",
            "type": "string"
          },
          "game_time_utc": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Game Time Utc"
          },
          "hitter_adjustment": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hitter Adjustment"
          },
          "home_team": {
            "$ref": "#/components/schemas/TeamBrief"
          },
          "humidity_percent": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Humidity Percent"
          },
          "observed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observed At"
          },
          "pitcher_adjustment": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pitcher Adjustment"
          },
          "precipitation_probability": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Precipitation Probability"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "temperature_f": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature F"
          },
          "venue": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VenueBrief"
              },
              {
                "type": "null"
              }
            ]
          },
          "wind_direction_degrees": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wind Direction Degrees"
          },
          "wind_out_mph": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wind Out Mph"
          },
          "wind_speed_mph": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wind Speed Mph"
          }
        },
        "required": [
          "game_id",
          "game_date",
          "away_team",
          "home_team",
          "available"
        ],
        "title": "WeatherData",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "All Rise Analytics API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/games": {
      "get": {
        "operationId": "games_api_v1_games_get",
        "parameters": [
          {
            "in": "query",
            "name": "date",
            "required": true,
            "schema": {
              "format": "date",
              "title": "Date",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "team",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 5,
                  "minLength": 2,
                  "pattern": "^[A-Za-z0-9]+$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Team"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 80,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_list_GameData__"
                }
              }
            },
            "description": "Successful Response"
          },
          "304": {
            "description": "The persisted data version has not changed"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Games",
        "tags": [
          "slate"
        ]
      }
    },
    "/api/v1/games/{game_id}": {
      "get": {
        "operationId": "game_api_v1_games__game_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "game_id",
            "required": true,
            "schema": {
              "maxLength": 80,
              "minLength": 3,
              "pattern": "^[A-Za-z0-9_.:-]+$",
              "title": "Game Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_GameData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "304": {
            "description": "The persisted data version has not changed"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Game",
        "tags": [
          "slate"
        ]
      }
    },
    "/api/v1/games/{game_id}/weather": {
      "get": {
        "operationId": "game_weather_api_v1_games__game_id__weather_get",
        "parameters": [
          {
            "in": "path",
            "name": "game_id",
            "required": true,
            "schema": {
              "maxLength": 80,
              "minLength": 3,
              "pattern": "^[A-Za-z0-9_.:-]+$",
              "title": "Game Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_WeatherData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "304": {
            "description": "The persisted data version has not changed"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Game Weather",
        "tags": [
          "slate"
        ]
      }
    },
    "/api/v1/weather": {
      "get": {
        "operationId": "weather_api_v1_weather_get",
        "parameters": [
          {
            "in": "query",
            "name": "date",
            "required": true,
            "schema": {
              "format": "date",
              "title": "Date",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "game_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 80,
                  "minLength": 3,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Game Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_list_WeatherData__"
                }
              }
            },
            "description": "Successful Response"
          },
          "304": {
            "description": "The persisted data version has not changed"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Weather",
        "tags": [
          "slate"
        ]
      }
    },
...
  }
}